Combine parameter estimates with parameter key

Combines model output parameter estimates with information in parameter key. Performs some formatting of this combined data.frame.

Expected input is a data.frame with parameter estimates, with the columns: parameter_names, estimate, stderr, random_effect_sd, random_effect_sdse, fixed, diag, shrinkage.

Some parameter_names have punctuation such as OMEGA(1,1). A new column is added without punctuation, such as OMEGA11.

Following this, parameter details from the parameter key are joined to the parameter estimates. A dplyr::inner_join is used so that only parameters in the model output are kept in the table. This was done so that, if your base and final model used the same structural THETAs and random parameters, the same parameter key could be used for both.

This join adds the following columns: abb (abbreviation), desc (parameter description), panel, trans (transformation).

With this information provided, a check is performed to determine whether parameters with special transformation rules were defined correctly. In addition, a series of TRUE/FALSE columns that will be used by subsequent functions.

defineParamTable(.estimates, .key, .ci = 95, .zed = NULL)

Arguments

.estimates

parameter estimates data.frame

.key

parameter key

.ci

specify 90 or 95 percent confidence interval (default 95%)

.zed

z-score for the specified confidence interval. Only needed for confidence intervals that are NOT 90 or 95 percent

See also