Format parameter table

Format parameter estimate values and output selected columns to be shown in the parameter table.

There are four main steps of this function:

When necessary, back transform parameters and their CIs, round parameters using pmtables::sig, and combine columns.

Format the THETA/OMEGA/SIGMA values to display as greek letters in Latex, with subscript numbers, and where necessary, the transformation applied to that parameter.

Determine which panel of the final table the parameter should be displayed in. This is informed by the panel argument you defined in your parameter key. Note that there are a finite number of options included by default (see below) but you can include additional panels as needed. Panel types include:

Residual variance

Interindividual covariance parameters

Interindividual variance parameters

Interoccasion variance parameters

Covariate effect parameters

Structural model parameters

Select columns for final tables.

formatParamTable(
  .df,
  .select_cols = c("type", "abb", "greek", "desc", "value", "ci", "shrinkage"),
  .prse = FALSE,
  .digit = getOption("mrgparamtab.dig"),
  .maxex = getOption("mrgparamtab.maxex")
)

Arguments

.df

parameter estimates with modifications ready for formatting

.select_cols

columns to select for output. Default selects "type", "abb", "greek", "desc", "value", "ci", "shrinkage". To return all columns, specify "all" for .select_cols.

.prse

output pRSE. Default is FALSE.

.digit

set significant digits for output (optional)

.maxex

set maxex for computation (optional)